feature: improved encoder script for best practices
